+++ python/lib/sift_client/client.py
@@ -48,6 +48,8 @@ from sift_client.resources import (
     UserDefinedFunctionsAPIAsync,
     UsersAPI,
     UsersAPIAsync,
+    WebhooksAPI,
+    WebhooksAPIAsync,
 )
 from sift_client.resources.access_control import AccessControlAPI, AccessControlAPIAsync
 from sift_client.transport import (
@@ -154,6 +156,9 @@ class SiftClient(
 
     users: UsersAPI
     """Instance of the Users API for making synchronous requests."""
+
+    webhooks: WebhooksAPI
+    """Instance of the Webhooks API for making synchronous requests."""
 
     data_export: DataExportAPI
     """Instance of the Data Export API for making synchronous requests."""
@@ -267,6 +272,7 @@ class SiftClient(
         self.test_results = TestResultsAPI(self)
         self.user_defined_functions = UserDefinedFunctionsAPI(self)
         self.users = UsersAPI(self)
+        self.webhooks = WebhooksAPI(self)
         self.data_export = DataExportAPI(self)
         self.data_import = DataImportAPI(self)
 
@@ -292,6 +298,7 @@ class SiftClient(
             test_results=TestResultsAPIAsync(self),
             user_defined_functions=UserDefinedFunctionsAPIAsync(self),
             users=UsersAPIAsync(self),
+            webhooks=WebhooksAPIAsync(self),
             data_export=DataExportAPIAsync(self),
             data_import=DataImportAPIAsync(self),
         )