+++ python/lib/sift_client/util/util.py
@@ -22,6 +22,7 @@ if TYPE_CHECKING:
         TestResultsAPIAsync,
         UserDefinedFunctionsAPIAsync,
         UsersAPIAsync,
+        WebhooksAPIAsync,
     )
     from sift_client.resources.access_control import AccessControlAPIAsync
 
@@ -79,6 +80,9 @@ class AsyncAPIs(NamedTuple):
 
     users: UsersAPIAsync
     """Instance of the Users API for making asynchronous requests."""
+
+    webhooks: WebhooksAPIAsync
+    """Instance of the Webhooks API for making asynchronous requests."""
 
     data_export: DataExportAPIAsync
     """Instance of the Data Export API for making asynchronous requests."""